Refactor the change shipping address and change card buttons to a more modular structure
