+++ modules/ppcp-axo-block/resources/js/components/Card/CardChangeButton.js
@@ -1,0 +1,18 @@
+import { createElement } from '@wordpress/element';
+
+const CardChangeButton = ( { onChangeButtonClick } ) =>
+	createElement(
+		'a',
+		{
+			className:
+				'wc-block-checkout-axo-block-card__edit wc-block-axo-change-link',
+			role: 'button',
+			onClick: ( event ) => {
+				event.preventDefault();
+				onChangeButtonClick();
+			},
+		},
+		'Choose a different card'
+	);
+
+export default CardChangeButton;
