+++ modules/ppcp-axo-block/resources/js/components/Card/index.js
@@ -1,0 +1,4 @@
+export { default as Card } from './Card';
+export { default as CardChangeButton } from './CardChangeButton';
+export { default as CardChangeButtonManager } from './CardChangeButtonManager';
+export { injectCardChangeButton, removeCardChangeButton } from './utils';
