+++ modules/ppcp-axo-block/resources/js/components/EmailButton/index.js
@@ -1,0 +1,6 @@
+export { default as EmailButton } from './EmailButton';
+export {
+	setupEmailFunctionality,
+	removeEmailFunctionality,
+	isEmailFunctionalitySetup,
+} from './utils';
