+++ modules/ppcp-axo-block/resources/js/components/Shipping/index.js
@@ -1,9 +1,6 @@
 export { default as ShippingChangeButton } from './ShippingChangeButton';
 export { default as ShippingChangeButtonManager } from './ShippingChangeButtonManager';
 export {
-	snapshotFields,
-	restoreOriginalFields,
-	populateWooFields,
 	injectShippingChangeButton,
 	removeShippingChangeButton,
 } from './utils';
