+++ modules/ppcp-axo-block/resources/js/components/Watermark/index.js
