--- modules/ppcp-axo-block/resources/js/components/email-button/index.js
@@ -1,15 +1,0 @@
-import { EmailButton } from './EmailButton';
-import {
-	setupEmailFunctionality,
-	removeEmailFunctionality,
-	isEmailFunctionalitySetup,
-} from './utils';
-
-export {
-	EmailButton,
-	setupEmailFunctionality,
-	removeEmailFunctionality,
-	isEmailFunctionalitySetup,
-};
-
-export default EmailButton;
