+++ modules/ppcp-axo-block/resources/js/hooks/useCustomerData.js
@@ -1,6 +1,5 @@
 import { useCallback, useMemo } from '@wordpress/element';
 import { useDispatch, useSelect } from '@wordpress/data';
-import useFastlaneSdk from "./useFastlaneSdk";
 
 export const useCustomerData = () => {
 	const customerData = useSelect( ( select ) =>
